+++ subpackages/utils/source/odood/utils/addons/addon.d
@@ -26,6 +26,7 @@ private import odood.utils.addons.addon_manifest;
 final class OdooAddon {
     private immutable string _name;
     private immutable Path _path;
+    private immutable Path _manifest_path;
     private OdooAddonManifest _manifest;
 
     @disable this();
@@ -37,9 +38,10 @@ final class OdooAddon {
       *     path = Path to addon on filesystem
       **/
     this(in Path path) {
-        this._path = path.toAbsolute;
-        this._name = _path.baseName;
-        this._manifest = parseOdooManifest(getAddonManifestPath(_path).get);
+        _path = path.toAbsolute;
+        _name = _path.baseName;
+        _manifest_path = getAddonManifestPath(_path).get;
+        _manifest = parseOdooManifest(_manifest_path);
     }
 
     /// name of the addon
@@ -47,6 +49,9 @@ final class OdooAddon {
 
     /// path to the addon on filesystem
     auto path() const => _path;
+
+    /// path to addon's manifest
+    auto manifest_path() const => _manifest_path;
 
     /// module manifest
     auto manifest() const => _manifest;